requirements.txt
# ===================================
# AI就活支援Webアプリ - 必要パッケージ一覧
# ===================================

# =============================================================================
# CORE FRAMEWORK & WEB
# =============================================================================
Django==5.1.2                       # Webフレームワーク
python-dotenv==1.1.0                # .envファイル読み込み

# =============================================================================
# DATABASE & ORM
# =============================================================================
psycopg2-binary>=2.9.7              # PostgreSQL adapter

# =============================================================================
# MACHINE LEARNING & DATA SCIENCE
# =============================================================================
pandas==2.1.3                       # データ分析・操作
numpy==1.26.3                       # 数値計算
scikit-learn==1.4.2                 # 機械学習ライブラリ
umap-learn==0.5.3                   # UMAP次元削減
hdbscan==0.8.33                     # 密度ベースクラスタリング
statsmodels==0.14.2                 # ARIMAモデルなど統計モデリング

# =============================================================================
# VISUALIZATION
# =============================================================================
matplotlib==3.8.4                   # グラフ描画
japanize-matplotlib==1.1.3          # 日本語フォント対応

# =============================================================================
# AI & LLM INTEGRATION
# =============================================================================
google-generativeai==0.8.5          # Gemini API統合
google-ai-generativelanguage==0.6.15
google-api-core>=2.25.0
google-auth>=2.40.0

# =============================================================================
# WEB SEARCH & INFORMATION RETRIEVAL
# =============================================================================
tavily-python==0.7.8                # Web検索API統合
requests==2.31.0                    # HTTPリクエスト（API呼び出し用）

# =============================================================================
# EDINET DATA FETCHER
# =============================================================================
beautifulsoup4>=4.12.0              # XBRL解析用HTMLパース
lxml>=4.9.0                         # XBRL解析用XMLパース

# =============================================================================
# UTILITY & SUPPORT
# =============================================================================
Pillow>=10.0.0                      # 画像処理（matplotlib等で必要）

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
pytest>=7.0.0                       # テストフレームワーク
pytest-django>=4.5.0               # Django用テスト拡張

# =============================================================================
# PRODUCTION & DEPLOYMENT
# =============================================================================
gunicorn>=21.0.0                    # WSGIサーバー
whitenoise>=6.5.0                   # 静的ファイル配信
django-environ>=0.11.0              # 環境設定管理

# =============================================================================
# NOTES
# =============================================================================
# 1. 本プロジェクトで使用されている主要な技術スタック:
#    - Django (Web Framework)
#    - scikit-learn (機械学習)
#    - pandas/numpy (データ処理)
#    - matplotlib (可視化)
#    - Gemini API (AI分析)
#    - Tavily API (Web検索)
#    - EDINET-DATAFETCHER (財務データ取得)
#
# 2. インストール方法:
#    pip install -r requirements.txt
#
# 3. 仮想環境推奨:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
#
# 4. 環境変数設定が必要:
#    - GEMINI_API_KEY
#    - TAVILY_API_KEY
#    - DJANGO_SECRET_KEY
#    - EDINET_API_KEY (データベース構築時)
#
# 5. データベース構築:
#    EDINET-DATAFETCHERを使用してEDINET APIから財務データを取得
#    詳細は README.md の「データベースの構築」セクションを参照
